requirements.txt
openai
langchain
streamlit
streamlit-chat
tiktoken
python-dotenv
pypdf
pinecone
pydantic
langchain-community
langchain-pinecone
langchainhub
langchain-openai
mysql-connector-python
beautifulsoup4
openpyxl
Levenshtein
duckduckgo-search
langchain-experimental